Init import/export refactoring
