+++ src/core/index.ts
@@ -1,3 +1,4 @@
-import { PieceMoveType } from "core/move_type.js";
-
-export { PieceMoveType }
+export { Board } from "./Board.js";
+export { CheckManager } from "./CheckManager.js";
+export { PromotionManager } from "./PromotionManager.js";
+export { PieceMoveType } from "./PieceMoveType.js";
