+++ src/pieces/movements/index.ts
@@ -1,0 +1,3 @@
+export { PieceMovement } from "./PieceMovement.js";
+export { CloseRangeMovement } from "./CloseRangeMovement.js";
+export { LongRangeMovement } from "./LongRangeMovement.js";
